requirements.txt
discord.py
openai
shodan
python-dotenv
